+++ pepe-tg/src/actions/index.ts
@@ -1,5 +1,6 @@
 export { fakeRaresCardAction } from "./fakeRaresCard";
 export { fakeCommonsCardAction } from "./fakeCommonsCard";
+export { rarePepesCardAction } from "./rarePepesCard";
 export { educateNewcomerAction } from "./educateNewcomer";
 export { startCommand, helpCommand } from "./basicCommands";
 export { loreCommand } from "./loreCommand";